tts rerender fixes
