+++ src/components/a1/messages/group.tsx
@@ -9,7 +9,6 @@ export const MessageGroup = ({
   messageRole,
   contentToCopy,
   contentToSpeak,
-  showTts,
   messageId,
   onEdit,
   onBranch,
@@ -18,7 +17,6 @@ export const MessageGroup = ({
   messageRole: UIMessage["role"];
   contentToCopy: string;
   contentToSpeak: string;
-  showTts: boolean;
   messageId: UIMessage["id"];
   onEdit?: () => void;
   onBranch?: () => void;
@@ -41,7 +39,6 @@ export const MessageGroup = ({
       <MessageActionRow
         contentToCopy={contentToCopy}
         contentToSpeak={contentToSpeak}
-        showTts={showTts}
         messageRole={messageRole}
         messageId={messageId}
         onEdit={onEdit}
